requirements.txt
# requirements.txt - Version compatible Windows
Django==4.2.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
Pillow==12.1.0 # Version plus ancienne mais stable
python-dotenv==1.0.0
whitenoise==6.6.0
dj-database-url==2.1.0
django-crispy-forms==2.1
crispy-bootstrap5>=2024.2
